improve performance of Java implementation; now it works real-time with small preview sizes

+++ FaceDetection/Core/CameraPreviewEventArgs.cs
@@ -3,7 +3,7 @@ namespace FaceDetection
 {
   public class CameraPreviewEventArgs : EventArgs
   {
-    public int FrameOrder { get; set; }
+    public long FrameOrder { get; set; }
     public byte[] FrameData { get; set; }
     public int Width { get; set; }
     public int Height { get; set; }
